Cleaning up.
